Revert "fix(InitializePipeline): Delete existing violations and TXCFileAttrib…"
This reverts commit 47d09d2f9e23862609104594aff7170f46e6cfe6.
