+++ src/timetables_etl/initialize_pipeline/app/initialize_pipeline.py
@@ -3,26 +3,28 @@ Lambda: InitializePipeline
 """
 
 from typing import Any
+from uuid import uuid4
 
 from aws_lambda_powertools.metrics import MetricUnit
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from common_layer.aws import configure_metrics
 from common_layer.database.client import SqlDB
-from common_layer.database.models import DatasetETLTaskResult
-from common_layer.database.repos import ETLTaskResultRepo
+from common_layer.database.models import (
+    DatasetETLTaskResult,
+    OrganisationDatasetRevision,
+    TaskState,
+)
+from common_layer.database.repos import (
+    ETLTaskResultRepo,
+    OrganisationDatasetRevisionRepo,
+)
 from common_layer.dynamodb.client.cache import DynamoDBCache
 from common_layer.dynamodb.data_manager import FileProcessingDataManager
+from common_layer.enums import FeedStatus
+from common_layer.exceptions.pipeline_exceptions import PipelineException
 from common_layer.json_logging import configure_logging
 from pydantic import BaseModel
 from structlog.stdlib import get_logger
-
-from .db_operations import (
-    create_task_result,
-    delete_existing_txc_file_attributes,
-    delete_existing_validation_violations,
-    get_and_validate_revision,
-    update_revision_status,
-)
 
 metrics = configure_metrics()
 logger = get_logger()
@@ -34,6 +36,48 @@ class InitializePipelineEvent(BaseModel):
     """
 
     DatasetRevisionId: int
+
+
+def get_and_validate_revision(
+    db: SqlDB, revision_id: int
+) -> OrganisationDatasetRevision:
+    """
+    Retrieves and validates the existence of a dataset revision.
+    """
+    revision_repo = OrganisationDatasetRevisionRepo(db)
+    revision = revision_repo.get_by_id(revision_id)
+    if revision is None:
+        raise PipelineException(f"DatasetRevision with id {revision_id} not found.")
+    return revision
+
+
+def update_revision_status(db: SqlDB, revision: OrganisationDatasetRevision) -> None:
+    """
+    Updates the revision status to indexing.
+    """
+    logger.debug(
+        "Setting OrganisationDatasetRevision Status to indexing",
+        dataset_revision_id=revision.id,
+    )
+    revision_repo = OrganisationDatasetRevisionRepo(db)
+    revision.status = FeedStatus.INDEXING.value
+    revision_repo.update(revision)
+
+
+def create_task_result(db: SqlDB, revision_id: int) -> DatasetETLTaskResult:
+    """
+    Creates a new ETL task result entry.
+    Returns:
+        ID of the created task result
+    """
+    task_result_repo = ETLTaskResultRepo(db)
+    task_result = DatasetETLTaskResult(
+        revision_id=revision_id,
+        status=TaskState.STARTED,
+        task_id=str(uuid4()),
+    )
+    created_task_result = task_result_repo.insert(task_result)
+    return created_task_result
 
 
 def initialize_pipeline(
@@ -50,8 +94,6 @@ def initialize_pipeline(
     revision = get_and_validate_revision(db, event.DatasetRevisionId)
 
     update_revision_status(db, revision)
-    delete_existing_validation_violations(db, revision.id)
-    delete_existing_txc_file_attributes(db, revision.id)
 
     task_result = create_task_result(db, revision.id)
